+++ src/lib.rs
@@ -10,6 +10,7 @@
 pub mod action;
 pub mod api;
 pub mod app;
+pub mod bee_supervisor;
 pub mod cli;
 pub mod components;
 pub mod config;